requirements.txt
streamlit>=1.32.0
langchain>=0.3.0
langchain-openai>=0.0.8
openai>=1.14.3
python-dotenv>=1.0.1
tiktoken>=0.6.0